+++ src/main/java/boundary/panel/pembelian/CheckoutPanel.java
@@ -1,21 +1,23 @@
-package PembelianUI;
+package boundary.panel.pembelian;
 
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
 
+import Util.Colors;
 import Util.RupiahConverter;
+import boundary.widget.RoundedPanel;
 
 import java.awt.*;
 
-public class TerimakasihPanel extends JPanel {
+public class CheckoutPanel extends JPanel {
   private float sub = 0f, discount = 0f, tax = 0f, total = 0f;
-  private String memberName;
+  private String[] memberNameList = { "", "Kim", "Jisoo", "Unnie" };
 
   // UI Components
   private JButton exitButton = new JButton();
-  private JLabel terimakasihLabel = new JLabel("Terimakasih!");
+  private JLabel checkoutLabel = new JLabel("Cek Keluar");
 
-  private JLabel billHeader = new JLabel("Fixed Bill");
+  private JLabel detailHeader = new JLabel("Detail Pesanan");
   private JPanel summaryTextPanel = new JPanel(new GridBagLayout());
 
   private JPanel subTextContainer = new JPanel(new BorderLayout());
@@ -43,16 +45,19 @@ public class TerimakasihPanel extends JPanel {
   private JLabel totalValue = new JLabel(RupiahConverter.convert(total));
 
   private JLabel memberHeader = new JLabel("Member");
-  private JLabel memberText = new JLabel("Transaksi ini dimiliki customer non-Member.");
-  private RoundedPanel tambahButtonContainer = new RoundedPanel(36, new Color(74, 107, 222), false, Color.WHITE, 0);
-  private JButton tambahButton = new JButton("+ Tambah");
-  private JLabel memberNameText = new JLabel("");
+  private JComboBox<String> memberDropdown = new JComboBox<>(memberNameList);
 
-  private RoundedPanel selesaiButtonContainer = new RoundedPanel(24, new Color(74, 107, 222), false, Color.WHITE, 0);
-  private JButton selesaiButton = new JButton("Selesai");
+  private RoundedPanel unduhButtonContainer = new RoundedPanel(10, Color.WHITE, true, new Color(82, 117, 226), 2);
+  private JButton unduhButton = new JButton("Unduh Transaksi");
 
-  public TerimakasihPanel(String memberName) {
-    this.memberName = memberName;
+  private RoundedPanel batalButtonContainer = new RoundedPanel(10, Color.WHITE, true, Colors.RED.getColor(), 2);
+  private JButton batalButton = new JButton("Batal");
+
+  private RoundedPanel orderButtonContainer = new RoundedPanel(10, Colors.BUTTON_BLUE.getColor(), false, Color.WHITE,
+      0);
+  private JButton orderButton = new JButton("Place Order");
+
+  public CheckoutPanel() {
     this.initializeUI();
   }
 
@@ -61,7 +66,7 @@ public class TerimakasihPanel extends JPanel {
     this.setBackground(Color.WHITE);
 
     ImageIcon buttonImage = new ImageIcon(
-        "/home/rma1403/Documents/Programming/kuliah/Tubes-2-OOP-O08/src/main/java/img/left-arrow.png");
+        "/home/rma1403/Documents/Programming/kuliah/Tubes-2-OOP-O08/src/main/resources/assets/icon/left-arrow.png");
     ImageIcon buttonImageScaled = new ImageIcon(
         buttonImage.getImage().getScaledInstance(18, 23, java.awt.Image.SCALE_SMOOTH));
     exitButton.setIcon(buttonImageScaled);
@@ -72,15 +77,15 @@ public class TerimakasihPanel extends JPanel {
     exitButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
     this.add(exitButton);
 
-    terimakasihLabel.setBounds(122, 66, 250, 40);
-    terimakasihLabel.setFont(new Font("Inter", Font.BOLD, 32));
-    terimakasihLabel.setForeground(new Color(229, 151, 0));
-    this.add(terimakasihLabel);
+    checkoutLabel.setBounds(122, 66, 180, 40);
+    checkoutLabel.setFont(new Font("Inter", Font.BOLD, 32));
+    checkoutLabel.setForeground(Colors.ORANGE.getColor());
+    this.add(checkoutLabel);
 
-    billHeader.setForeground(new Color(36, 60, 148));
-    billHeader.setFont(new Font("Inter", Font.BOLD, 24));
-    billHeader.setBounds(105, 151, 193, 30);
-    this.add(billHeader);
+    detailHeader.setForeground(Colors.DARK_BLUE.getColor());
+    detailHeader.setFont(new Font("Inter", Font.BOLD, 24));
+    detailHeader.setBounds(105, 151, 193, 30);
+    this.add(detailHeader);
 
     summaryTextPanel.setBackground(Color.WHITE);
     summaryTextPanel.setBounds(106, 201, 355, 135);
@@ -92,7 +97,7 @@ public class TerimakasihPanel extends JPanel {
     subTextContainer.setBackground(Color.WHITE);
     subTextContainer.setBorder(new EmptyBorder(0, 0, 20, 0));
     subText.setPreferredSize(new Dimension(80, 18));
-    subText.setForeground(new Color(36, 60, 148));
+    subText.setForeground(Colors.DARK_BLUE.getColor());
     subText.setFont(summaryFont);
     subTextContainer.add(subText, BorderLayout.CENTER);
     summaryTextPanel.add(subTextContainer, c);
@@ -102,7 +107,7 @@ public class TerimakasihPanel extends JPanel {
     subValueContainer.setBackground(Color.WHITE);
     subValueContainer.setBorder(new EmptyBorder(0, 0, 20, 0));
     subValue.setPreferredSize(new Dimension(266, 18));
-    subValue.setForeground(new Color(36, 60, 148));
+    subValue.setForeground(Colors.DARK_BLUE.getColor());
     subValue.setFont(summaryFont);
     subValue.setHorizontalAlignment(SwingConstants.RIGHT);
     subValueContainer.add(subValue, BorderLayout.CENTER);
@@ -113,7 +118,7 @@ public class TerimakasihPanel extends JPanel {
     discountTextContainer.setBackground(Color.WHITE);
     discountTextContainer.setBorder(new EmptyBorder(0, 0, 20, 0));
     discountText.setPreferredSize(new Dimension(80, 18));
-    discountText.setForeground(new Color(36, 60, 148));
+    discountText.setForeground(Colors.DARK_BLUE.getColor());
     discountText.setFont(summaryFont);
     discountTextContainer.add(discountText, BorderLayout.CENTER);
     summaryTextPanel.add(discountTextContainer, c);
@@ -123,7 +128,7 @@ public class TerimakasihPanel extends JPanel {
     discountValueContainer.setBackground(Color.WHITE);
     discountValueContainer.setBorder(new EmptyBorder(0, 0, 20, 0));
     discountValue.setPreferredSize(new Dimension(266, 18));
-    discountValue.setForeground(new Color(36, 60, 148));
+    discountValue.setForeground(Colors.DARK_BLUE.getColor());
     discountValue.setFont(summaryFont);
     discountValue.setHorizontalAlignment(SwingConstants.RIGHT);
     discountValueContainer.add(discountValue, BorderLayout.CENTER);
@@ -134,7 +139,7 @@ public class TerimakasihPanel extends JPanel {
     taxTextContainer.setBackground(Color.WHITE);
     taxTextContainer.setBorder(new EmptyBorder(0, 0, 20, 0));
     taxText.setPreferredSize(new Dimension(80, 18));
-    taxText.setForeground(new Color(36, 60, 148));
+    taxText.setForeground(Colors.DARK_BLUE.getColor());
     taxText.setFont(summaryFont);
     taxTextContainer.add(taxText, BorderLayout.CENTER);
     summaryTextPanel.add(taxTextContainer, c);
@@ -144,7 +149,7 @@ public class TerimakasihPanel extends JPanel {
     taxValueContainer.setBackground(Color.WHITE);
     taxValueContainer.setBorder(new EmptyBorder(0, 0, 20, 0));
     taxValue.setPreferredSize(new Dimension(266, 18));
-    taxValue.setForeground(new Color(36, 60, 148));
+    taxValue.setForeground(Colors.DARK_BLUE.getColor());
     taxValue.setFont(summaryFont);
     taxValue.setHorizontalAlignment(SwingConstants.RIGHT);
     taxValueContainer.add(taxValue, BorderLayout.CENTER);
@@ -154,7 +159,7 @@ public class TerimakasihPanel extends JPanel {
     c.gridy = 3;
     totalTextContainer.setBackground(Color.WHITE);
     totalText.setPreferredSize(new Dimension(80, 18));
-    totalText.setForeground(new Color(36, 60, 148));
+    totalText.setForeground(Colors.DARK_BLUE.getColor());
     totalText.setFont(summaryFont);
     totalTextContainer.add(totalText, BorderLayout.CENTER);
     summaryTextPanel.add(totalTextContainer, c);
@@ -163,7 +168,7 @@ public class TerimakasihPanel extends JPanel {
     c.gridy = 3;
     totalValueContainer.setBackground(Color.WHITE);
     totalValue.setPreferredSize(new Dimension(266, 18));
-    totalValue.setForeground(new Color(36, 60, 148));
+    totalValue.setForeground(Colors.DARK_BLUE.getColor());
     totalValue.setFont(summaryFont);
     totalValue.setHorizontalAlignment(SwingConstants.RIGHT);
     totalValueContainer.add(totalValue, BorderLayout.CENTER);
@@ -171,52 +176,53 @@ public class TerimakasihPanel extends JPanel {
 
     this.add(summaryTextPanel);
 
-    memberHeader.setBounds(105, 394, 125, 35);
+    memberHeader.setBounds(105, 384, 125, 35);
     memberHeader.setFont(new Font("Inter", Font.BOLD, 24));
-    memberHeader.setForeground(new Color(36, 60, 148));
+    memberHeader.setForeground(Colors.DARK_BLUE.getColor());
     this.add(memberHeader);
 
-    memberText.setBounds(105, 425, 350, 25);
-    memberText.setFont(new Font("Inter", Font.PLAIN, 15));
-    memberText.setForeground(new Color(194, 194, 194));
-    if(memberName == "") {
-      this.add(memberText);
-    }
+    memberDropdown.setBounds(105, 427, 373, 34);
+    memberDropdown.setFocusable(false);
+    memberDropdown.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
+    this.add(memberDropdown);
 
-    memberNameText.setBounds(105, 445, 350, 25);
-    memberNameText.setFont(new Font("Inter", Font.BOLD, 18));
-    memberNameText.setHorizontalAlignment(SwingConstants.CENTER);
-    memberNameText.setText(memberName);
-    if(memberName != "" ) {
-      this.add(memberNameText);
-    }
+    unduhButtonContainer.setLayout(null);
+    unduhButtonContainer.setBounds(105, 503, 373, 49);
+    this.add(unduhButtonContainer);
 
-    tambahButtonContainer.setLayout(null);
-    tambahButtonContainer.setBounds(105, 454, 350, 36);
-    if(memberName == "") {
-      this.add(tambahButtonContainer);
-    }
+    unduhButton.setBorder(null);
+    unduhButton.setBackground(Color.WHITE);
+    unduhButton.setFocusPainted(false);
+    unduhButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
+    unduhButton.setBounds(5, 2, 363, 45);
+    unduhButton.setForeground(new Color(82, 117, 225));
+    unduhButton.setFont(new Font("Inter", Font.BOLD, 18));
+    unduhButtonContainer.add(unduhButton);
 
-    tambahButton.setBorder(null);
-    tambahButton.setBackground(new Color(74, 107, 222));
-    tambahButton.setFocusPainted(false);
-    tambahButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
-    tambahButton.setBounds(12, 2, 326, 32);
-    tambahButton.setForeground(Color.WHITE);
-    tambahButton.setFont(new Font("Inter", Font.BOLD, 16));
-    tambahButtonContainer.add(tambahButton);
+    batalButtonContainer.setLayout(null);
+    batalButtonContainer.setBounds(104, 569, 133, 47);
+    this.add(batalButtonContainer);
 
-    selesaiButtonContainer.setLayout(null);
-    selesaiButtonContainer.setBounds(180, 527, 205, 43);
-    this.add(selesaiButtonContainer);
+    batalButton.setBorder(null);
+    batalButton.setBackground(Color.WHITE);
+    batalButton.setFocusPainted(false);
+    batalButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
+    batalButton.setBounds(5, 2, 123, 43);
+    batalButton.setForeground(Colors.RED.getColor());
+    batalButton.setFont(new Font("Inter", Font.BOLD, 18));
+    batalButtonContainer.add(batalButton);
 
-    selesaiButton.setBorder(null);
-    selesaiButton.setBackground(new Color(74, 107, 222));
-    selesaiButton.setFocusPainted(false);
-    selesaiButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
-    selesaiButton.setBounds(9, 2, 187, 39);
-    selesaiButton.setForeground(Color.WHITE);
-    selesaiButton.setFont(new Font("Inter", Font.PLAIN, 18));
-    selesaiButtonContainer.add(selesaiButton);
+    orderButtonContainer.setLayout(null);
+    orderButtonContainer.setBounds(273, 569, 205, 47);
+    this.add(orderButtonContainer);
+
+    orderButton.setBorder(null);
+    orderButton.setBackground(Colors.BUTTON_BLUE.getColor());
+    orderButton.setFocusPainted(false);
+    orderButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
+    orderButton.setBounds(5, 2, 195, 43);
+    orderButton.setForeground(Color.WHITE);
+    orderButton.setFont(new Font("Inter", Font.BOLD, 18));
+    orderButtonContainer.add(orderButton);
   }
 }
