+++ src/main/java/boundary/panel/pembelian/CreateMemberPanel.java
@@ -1,7 +1,11 @@
-package PembelianUI;
+package boundary.panel.pembelian;
 
 import java.awt.*;
 import javax.swing.*;
+
+import Util.Colors;
+import boundary.widget.HintTextField;
+import boundary.widget.RoundedPanel;
 
 public class CreateMemberPanel extends JPanel {
   // UI Components
@@ -9,11 +13,11 @@ public class CreateMemberPanel extends JPanel {
   private JLabel checkoutLabel = new JLabel("Data Member Baru");
 
   private JLabel namaTextLabel = new JLabel("Nama");
-  private RoundedPanel namaFieldContainer = new RoundedPanel(40, Color.WHITE, true, new Color(36, 60, 148), 1);
+  private RoundedPanel namaFieldContainer = new RoundedPanel(40, Color.WHITE, true, Colors.DARK_BLUE.getColor(), 1);
   private JTextField namaField = new HintTextField("Masukkan nama pengguna...", 1);
 
   private JLabel teleponTextLabel = new JLabel("Nomor Telepon");
-  private RoundedPanel teleponFieldContainer = new RoundedPanel(40, Color.WHITE, true, new Color(36, 60, 148), 1);
+  private RoundedPanel teleponFieldContainer = new RoundedPanel(40, Color.WHITE, true, Colors.DARK_BLUE.getColor(), 1);
   private JTextField teleponField = new HintTextField("Masukkan nomor telepon...", 1);
 
   private JLabel memberTextLabel = new JLabel("Jenis Member");
@@ -21,7 +25,8 @@ public class CreateMemberPanel extends JPanel {
   private JRadioButton vipOption = new JRadioButton("VIP");
   ButtonGroup memberGroup = new ButtonGroup();
 
-  private RoundedPanel createButtonContainer = new RoundedPanel(24, new Color(74, 107, 222), false, Color.WHITE, 0);
+  private RoundedPanel createButtonContainer = new RoundedPanel(24, Colors.BUTTON_BLUE.getColor(), false, Color.WHITE,
+      0);
   private JButton createButton = new JButton("+ Create Member");
 
   public CreateMemberPanel() {
@@ -33,7 +38,7 @@ public class CreateMemberPanel extends JPanel {
     this.setBackground(Color.WHITE);
 
     ImageIcon buttonImage = new ImageIcon(
-        "/home/rma1403/Documents/Programming/kuliah/Tubes-2-OOP-O08/src/main/java/img/left-arrow.png");
+        "/home/rma1403/Documents/Programming/kuliah/Tubes-2-OOP-O08/src/main/resources/assets/icon/left-arrow.png");
     ImageIcon buttonImageScaled = new ImageIcon(
         buttonImage.getImage().getScaledInstance(18, 23, java.awt.Image.SCALE_SMOOTH));
     exitButton.setIcon(buttonImageScaled);
@@ -46,12 +51,12 @@ public class CreateMemberPanel extends JPanel {
 
     checkoutLabel.setBounds(122, 66, 330, 40);
     checkoutLabel.setFont(new Font("Inter", Font.BOLD, 32));
-    checkoutLabel.setForeground(new Color(229, 151, 0));
+    checkoutLabel.setForeground(Colors.ORANGE.getColor());
     this.add(checkoutLabel);
 
     namaTextLabel.setBounds(119, 151, 140, 24);
     namaTextLabel.setFont(new Font("Inter", Font.BOLD, 15));
-    namaTextLabel.setForeground(new Color(36, 60, 148));
+    namaTextLabel.setForeground(Colors.DARK_BLUE.getColor());
     this.add(namaTextLabel);
 
     namaFieldContainer.setLayout(null);
@@ -60,7 +65,7 @@ public class CreateMemberPanel extends JPanel {
 
     namaField.setFont(new Font("Inter", Font.PLAIN, 17));
     namaField.setBackground(Color.WHITE);
-    namaField.setForeground(new Color(36, 60, 148));
+    namaField.setForeground(Colors.DARK_BLUE.getColor());
     namaField.setCursor(Cursor.getPredefinedCursor(Cursor.TEXT_CURSOR));
     namaField.setBorder(null);
     namaField.setBounds(15, 7, 350, 25);
@@ -68,7 +73,7 @@ public class CreateMemberPanel extends JPanel {
 
     teleponTextLabel.setBounds(119, 249, 140, 24);
     teleponTextLabel.setFont(new Font("Inter", Font.BOLD, 15));
-    teleponTextLabel.setForeground(new Color(36, 60, 148));
+    teleponTextLabel.setForeground(Colors.DARK_BLUE.getColor());
     this.add(teleponTextLabel);
 
     teleponFieldContainer.setLayout(null);
@@ -77,7 +82,7 @@ public class CreateMemberPanel extends JPanel {
 
     teleponField.setFont(new Font("Inter", Font.PLAIN, 17));
     teleponField.setBackground(Color.WHITE);
-    teleponField.setForeground(new Color(36, 60, 148));
+    teleponField.setForeground(Colors.DARK_BLUE.getColor());
     teleponField.setCursor(Cursor.getPredefinedCursor(Cursor.TEXT_CURSOR));
     teleponField.setBorder(null);
     teleponField.setBounds(15, 7, 350, 25);
@@ -85,12 +90,12 @@ public class CreateMemberPanel extends JPanel {
 
     memberTextLabel.setBounds(119, 353, 140, 24);
     memberTextLabel.setFont(new Font("Inter", Font.BOLD, 15));
-    memberTextLabel.setForeground(new Color(36, 60, 148));
+    memberTextLabel.setForeground(Colors.DARK_BLUE.getColor());
     this.add(memberTextLabel);
 
     memberOption.setBounds(125, 378, 183, 32);
     memberOption.setFont(new Font("Inter", Font.BOLD, 15));
-    memberOption.setForeground(new Color(36, 60, 148));
+    memberOption.setForeground(Colors.DARK_BLUE.getColor());
     memberOption.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
     memberOption.setFocusPainted(false);
     memberOption.setOpaque(false);
@@ -98,7 +103,7 @@ public class CreateMemberPanel extends JPanel {
 
     vipOption.setBounds(125, 407, 183, 32);
     vipOption.setFont(new Font("Inter", Font.BOLD, 15));
-    vipOption.setForeground(new Color(36, 60, 148));
+    vipOption.setForeground(Colors.DARK_BLUE.getColor());
     vipOption.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
     vipOption.setFocusPainted(false);
     vipOption.setOpaque(false);
@@ -112,7 +117,7 @@ public class CreateMemberPanel extends JPanel {
     this.add(createButtonContainer);
 
     createButton.setBorder(null);
-    createButton.setBackground(new Color(74, 107, 222));
+    createButton.setBackground(Colors.BUTTON_BLUE.getColor());
     createButton.setFocusPainted(false);
     createButton.setCursor(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR));
     createButton.setBounds(9, 2, 194, 41);
