+++ src/main/resources/assets/icon/cross.png
