+++ src/main/resources/assets/icon/left-arrow.png
