+++ src/main/resources/assets/image/beef-contoh.png
