+++ src/main/resources/assets/image/salad-egg.png
